Addressing Pete's concerns
Also reworked the example to use JSF and inject both services via
qualifiers.
