+++ cdi-injection/src/main/java/org/jboss/as/quickstarts/cdi/injection/SpanishTranslateService.java
@@ -1,6 +1,6 @@
-package org.jboss.as.quickstarts.helloworld.translate;
+package org.jboss.as.quickstarts.cdi.injection;
 
-import org.jboss.as.quickstarts.helloworld.translate.qualifier.Spanish;
+import org.jboss.as.quickstarts.cdi.injection.qualifier.Spanish;
 
 /**
  * A simple implementation for the Spanish language.
