+++ cdi-injection/src/main/java/org/jboss/as/quickstarts/cdi/injection/TranslateService.java
@@ -1,4 +1,4 @@
-package org.jboss.as.quickstarts.helloworld.translate;
+package org.jboss.as.quickstarts.cdi.injection;
 
 /**
  * A simple interface which will define translation phrases supported by the system.
