+++ cdi-injection/src/main/webapp/WEB-INF/beans.xml
