+++ cdi-injection/src/main/webapp/index.html
@@ -2,6 +2,6 @@
 
 <html>
 <head>
-<meta http-equiv="Refresh" content="0; URL=HelloWorld">
+<meta http-equiv="Refresh" content="0; URL=translate.jsf">
 </head>
 </html>
