+++ cdi-injection/src/main/webapp/translate.xhtml
@@ -1,0 +1,20 @@
+<!DOCTYPE html
+        PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+        "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+   xmlns:ui="http://java.sun.com/jsf/facelets"
+   xmlns:h="http://java.sun.com/jsf/html"
+   xmlns:f="http://java.sun.com/jsf/core">
+
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1" />
+    <title>Translation Service</title>
+</head>
+<body>
+    <div>
+        English Hello: #{translation.englishHello}
+        <br />
+        Spanish Hello: #{translation.spanishHello}
+    </div>
+</body>
+</html>
