--- helloworld-translate/README.md
@@ -1,13 +1,0 @@
-Helloworld Example
-==================
-
-This example demonstrates the use of *CDI 1.0 injection* and *Servlet 3* in JBoss AS 7.
-
-The example can be deployed using Maven from the command line or from Eclipse using
-JBoss Tools.
-
-To set up Maven or JBoss Tools in Eclipse, refer to the _Getting Started Guide_.
-
-To deploy to JBoss AS 7, start JBoss AS 7 and type `mvn package jboss-as:deploy`.
-The application is deployed to <http://localhost:8080/jboss-as-helloworld-translate>. You
-can read more details in the _Getting Started Guide_.
