--- helloworld-translate/src/main/java/org/jboss/as/quickstarts/helloworld/translate/EnglishTranslateService.java
@@ -1,39 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat, Inc., and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the &quot;License&quot;);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an &quot;AS IS&quot; BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.as.quickstarts.helloworld.translate;
-
-import org.jboss.as.quickstarts.helloworld.translate.qualifier.English;
-
-/**
- * A simple implementation for the English language.
- *
- * The {@link English} qualifier tells CDI this is a special instance of
- * the {@link TranslateService}.
- *
- * @author Jason Porter
- *
- */
-@English
-public class EnglishTranslateService implements TranslateService {
-
-   @Override
-   public String hello() {
-     return "Hello";
-   }
-
-}
