--- helloworld-translate/src/main/java/org/jboss/as/quickstarts/helloworld/translate/qualifier/English.java
@@ -1,41 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2011, Red Hat, Inc., and individual contributors
- * by the @authors tag. See the copyright.txt in the distribution for a
- * full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the &quot;License&quot;);
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an &quot;AS IS&quot; BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jboss.as.quickstarts.helloworld.translate.qualifier;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.inject.Qualifier;
-
-/**
- * Qualifier used to help determine the intended injection object.
- *
- * This particular qualifier will be used to specify types which should be used for the English language.
- *
- * @author Jason Porter
- * @see Spanish
- */
-@Qualifier
-@Target({ElementType.TYPE, ElementType.FIELD, ElementType.PARAMETER})
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-public @interface English {
-}
